Cata\[Fly][A - Quest] Twilight Highlands 84-85 [Kick].xml
Fixed an issue during quest "He's Holding Out on Us" where toon is not able to navigate up the spiral stairs in SI:7. MyCTM was used to accomplish this before, but for some reason it failed. Changing the QB to RunLikeHell with UseCTM="True" made it work for some reason... fixed.